+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Enabled handling of redirect uri
 - Added third flow option login + permission
 - Added parameter to forece app2app flow
+- Added session persistence
 
 ### Changed
 - Renamed NetID to netID